chore: decompose large application modules
