+++ apps/web/src/components/search-dialog.tsx
@@ -1,19 +1,6 @@
-import {
-  lazy,
-  Suspense,
-  useCallback,
-  useId,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
-import type {
-  CSSProperties,
-  KeyboardEvent as ReactKeyboardEvent,
-  ReactNode,
-} from "react";
+import { useCallback, useMemo, useRef, useState } from "react";
+import type { CSSProperties, KeyboardEvent as ReactKeyboardEvent } from "react";
 
-import type { UseMutationResult } from "@tanstack/react-query";
 import {
   useInfiniteQuery,
   useMutation,
@@ -22,17 +9,7 @@ import {
 } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
 import { useVirtualizer } from "@tanstack/react-virtual";
-import {
-  FileTextIcon,
-  HistoryIcon,
-  LandmarkIcon,
-  LoaderIcon,
-  MessageSquareIcon,
-  MessagesSquareIcon,
-  PanelRightIcon,
-  UserIcon,
-  WandSparklesIcon,
-} from "lucide-react";
+import { LoaderIcon, PanelRightIcon, WandSparklesIcon } from "lucide-react";
 import { useDebouncedCallback } from "use-debounce";
 import { useTranslations } from "use-intl";
 
@@ -40,20 +17,16 @@ import {
   GLOBAL_SEARCH_RESULT_TYPES,
   type GlobalSearchResultType,
 } from "@stll/api-contract";
-import type { EntityKind, GlobalSearchHit } from "@stll/api/types";
-import { BidiText } from "@stll/ui/bidi-text";
+import type { GlobalSearchHit } from "@stll/api/types";
 import { Button } from "@stll/ui/button";
-import { Checkbox } from "@stll/ui/checkbox";
 import {
   Command,
   CommandDialog,
   CommandDialogPopup,
   CommandInput,
-  CommandItem,
   CommandList,
 } from "@stll/ui/command";
 import { DirectionalIcon } from "@stll/ui/directional-icon";
-import { Input } from "@stll/ui/input";
 import { Skeleton } from "@stll/ui/skeleton";
 import { stellaToast } from "@stll/ui/toast";
 import { contentDir } from "@stll/ui/use-content-dir";
@@ -61,9 +34,6 @@ import { useIsMobile } from "@stll/ui/use-mobile";
 import { cn } from "@stll/ui/utils";
 
 import { openEntityInInspector } from "@/components/chat/entity-open";
-import { DatePickerPopover } from "@/components/date-picker-popover";
-import { DocumentIcon } from "@/components/document-icon";
-import { MatterIcon } from "@/components/matter-icon";
 import { RenderStormRegion } from "@/components/render-storm-canary";
 import { SavedSearches } from "@/components/saved-searches";
 import {
@@ -71,16 +41,45 @@ import {
   type SavedSearchCriteria,
 } from "@/components/saved-searches.logic";
 import {
+  SearchColumnResizeHandle,
+  SearchFooterHint,
+  SearchFooterHintText,
+} from "@/components/search-dialog-controls";
+import {
+  FacetGroup,
+  SearchableFacetGroup,
+  TimeFacetGroup,
+} from "@/components/search-dialog-facets";
+import {
+  RecentFilePreviewPanel,
+  SearchPreviewPanel,
+} from "@/components/search-dialog-preview";
+import {
+  SearchRecents,
+  SearchResultItem,
+  SearchSummaryItem,
+} from "@/components/search-dialog-results";
+import {
   createDialogCloseActionQueue,
   getChatHitRoute,
   getEntityLocationRoute,
   getEntityWorkspaceRoute,
   getRecentFileRoute,
-  getRecentFilePreviewDateVisibility,
-  getRecentFilePreviewHit,
   resolveEntityDocumentRoute,
   toAskAIMessageHtml,
 } from "@/components/search-dialog.logic";
+import {
+  EMPTY_FACET_BUCKETS,
+  EMPTY_SEARCH_HITS,
+  EMPTY_SEARCH_PREVIEW_LOCATOR_CANDIDATES,
+  formatMimeTypeLabel,
+  initialSearchFilters,
+  isAvailableSearchKind,
+  isSearchKindOption,
+  KIND_TRANSLATION_KEYS,
+  mergeSelectedBuckets,
+  SEARCH_PREVIEW_COLUMN_CLASS_NAME,
+} from "@/components/search-dialog.shared";
 import {
   canShowSearchSummary,
   clearTime,
@@ -92,13 +91,7 @@ import {
   setPresetTime,
   toggleArrayMember,
 } from "@/components/search-filters.logic";
-import type {
-  SearchFilters,
-  TimeFilter,
-} from "@/components/search-filters.logic";
-import Tooltip from "@/components/tooltip";
-import { UserIdentity } from "@/components/user-avatar";
-import { EntityKindIcon } from "@/components/workspaces/entity-kind-icon";
+import type { SearchFilters } from "@/components/search-filters.logic";
 import { useChatUserContext } from "@/features/chat/hooks/use-chat-user-context";
 import { startNewThreadCommandHandoff } from "@/features/chat/lib/start-new-thread-command-handoff";
 import { invalidateGroupedChatThreads } from "@/features/chat/queries";
@@ -109,9 +102,8 @@ import {
   isPublicLawPreviewEnabled,
   usePublicLawPreviewEnabled,
 } from "@/hooks/use-public-law-preview";
-import { useFormatter, useLocale } from "@/i18n/formatting-context";
+import { useLocale } from "@/i18n/formatting-context";
 import { useI18nStore } from "@/i18n/i18n-store";
-import type { TranslationKey } from "@/i18n/types";
 import { useAnalytics } from "@/lib/analytics/provider";
 import { api } from "@/lib/api";
 import { useAuthenticatedUser } from "@/lib/authenticated-user-context";
@@ -119,25 +111,17 @@ import { createCaseLawDecisionRouteParams } from "@/lib/case-law-route";
 import { getChatSendMode } from "@/lib/chat-anonymized-store";
 import type { ChatThreadRef } from "@/lib/chat-thread-ref";
 import { createChatThreadId } from "@/lib/chat-thread-ref";
-import { DOCX_MIME, PDF_MIME } from "@/lib/consts";
 import { detached } from "@/lib/detached";
 import { unwrapEden } from "@/lib/errors/api";
 import { FILE_OPEN_TARGET, resolveFileOpenTarget } from "@/lib/file-open.logic";
 import { aiAvailabilityOptions } from "@/lib/organization/ai-config-queries";
 import { toSafeId } from "@/lib/safe-id";
 import {
-  searchFacetOptions,
   hasSearchQueryOrSelectiveFilter,
   recentFilePreviewFieldOptions,
   searchInfiniteOptions,
-  searchPreviewOptions,
-  TIME_PRESETS,
 } from "@/lib/search";
-import type {
-  SearchAISummaryParams,
-  SearchableFacet,
-  TimePreset,
-} from "@/lib/search";
+import type { SearchAISummaryParams, TimePreset } from "@/lib/search";
 import {
   readRecentFiles,
   readRecentSearches,
@@ -149,17 +133,8 @@ import type {
   RecentSearch,
   SearchRecentsScope,
 } from "@/lib/search-recents";
-import { searchTextQueryKey } from "@/lib/search-text";
 import {
-  getEmailSearchPreviewTarget,
   getFirstSearchHighlightText,
-  getNativeSearchDocumentPreviewTarget,
-  getSearchHighlightText,
-  getSearchPreviewRenderContent,
-  getSearchPreviewDate,
-  getSearchPreviewTarget,
-  normalizeSearchQuery,
-  selectAuthorizedSearchPreviewData,
   selectDisplayedSearchPreviewHit,
   selectSearchPreviewHit,
   shouldShowSearchPreview,
@@ -179,10 +154,6 @@ type CreateSearchSummaryChatVars = SearchAISummaryParams & {
   citations: SearchSummaryCitation[];
 };
 
-type SearchSummaryData = NonNullable<
-  Awaited<ReturnType<typeof api.search.summary.post>>["data"]
->;
-
 const DEBOUNCE_MS = 300;
 const PREVIEW_HIGHLIGHT_DEBOUNCE_MS = 180;
 const VIRTUAL_HIT_ESTIMATE_PX = 76;
@@ -200,137 +171,11 @@ const SEARCH_PREVIEW_MAX_WIDTH = 800;
 const SEARCH_PREVIEW_DEFAULT_WIDTH = 512;
 const SEARCH_RESULTS_MIN_WIDTH = 320;
 
-const SEARCH_PREVIEW_CONTENT_CLASS_NAME =
-  "text-foreground/90 [&_mark]:bg-highlight [&_mark]:text-highlight-foreground text-sm leading-6 whitespace-pre-wrap [&_mark]:font-medium";
-const SEARCH_PREVIEW_COLUMN_CLASS_NAME =
-  "hidden min-h-0 w-[var(--search-preview-w,min(44%,32rem))] min-w-72 shrink-0 flex-col overflow-hidden border-s md:flex";
-
-const NativeDocumentPreview = lazy(async () => {
-  const { SearchDocumentPreview } =
-    await import("@/components/search-document-preview");
-  return { default: SearchDocumentPreview };
-});
-
-const ChatPreview = lazy(async () => {
-  const { SearchChatPreview } =
-    await import("@/components/search-chat-preview");
-  return { default: SearchChatPreview };
-});
-
-// The same email HTML viewer the inspector renders, so an email previews as
-// a message, not as extracted text.
-const EmailPreview = lazy(async () => {
-  const { EmailHtmlViewer } =
-    await import("@/components/inspector/email-html-viewer");
-  return { default: EmailHtmlViewer };
-});
-
-const ignoreSearchPreviewAttachment = (): undefined => undefined;
-
-const NON_ENTITY_KIND_ICONS = {
-  contact: UserIcon,
-  "case-law": LandmarkIcon,
-  chat: MessagesSquareIcon,
-} as const satisfies Record<
-  Exclude<GlobalSearchResultType, EntityKind | "matter">,
-  typeof UserIcon
->;
-
-const KIND_TRANSLATION_KEYS = {
-  matter: "search.kinds.matter",
-  contact: "search.kinds.contact",
-  "case-law": "search.kinds.caseLaw",
-  document: "common.document",
-  folder: "search.kinds.folder",
-  task: "search.kinds.task",
-  message: "search.kinds.message",
-  link: "search.kinds.link",
-  chat: "search.kinds.chat",
-} as const satisfies Record<GlobalSearchResultType, TranslationKey>;
-
-const TIME_PRESET_TRANSLATION_KEYS = {
-  day: "search.updatedWithinOptions.day",
-  week: "search.updatedWithinOptions.week",
-  month: "search.updatedWithinOptions.month",
-  year: "search.updatedWithinOptions.year",
-} as const satisfies Record<TimePreset, TranslationKey>;
-
-const isSearchKindOption = (value: string): value is GlobalSearchResultType =>
-  GLOBAL_SEARCH_RESULT_TYPES.some((type) => type === value);
-
-const compactMeta = (parts: readonly (string | null | undefined)[]): string =>
-  parts
-    .flatMap((part) => {
-      const trimmed = part?.trim();
-      return trimmed ? [trimmed] : [];
-    })
-    .join(" · ");
-
-const isAvailableSearchKind = (
-  type: GlobalSearchResultType,
-  includePublicLaw: boolean,
-): boolean => type !== "case-law" || includePublicLaw;
-
-const formatMimeTypeLabel = (mimeType: string): string => {
-  if (mimeType === "application/pdf") {
-    return "PDF";
-  }
-  if (
-    mimeType ===
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document"
-  ) {
-    return "DOCX";
-  }
-  if (
-    mimeType ===
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
-  ) {
-    return "XLSX";
-  }
-  if (mimeType.startsWith("image/")) {
-    return mimeType.replace("image/", "").toUpperCase();
-  }
-  if (mimeType.startsWith("text/")) {
-    return mimeType.replace("text/", "").toUpperCase();
-  }
-  return mimeType;
-};
-
-const isoToDateInputValue = (iso: string): string => iso.slice(0, 10);
-
-const dateInputToIsoStart = (value: string): string =>
-  new Date(`${value}T00:00:00.000Z`).toISOString();
-
-const dateInputToIsoEnd = (value: string): string =>
-  new Date(`${value}T23:59:59.999Z`).toISOString();
-
-const mergeSelectedBuckets = (
-  buckets: { value: string; label?: string; count: number }[],
-  selected: string[],
-  getLabel: (value: string) => string,
-): { value: string; label?: string; count: number }[] => {
-  const present = new Set(buckets.map((bucket) => bucket.value));
-  const missing = selected.flatMap((value) =>
-    present.has(value) ? [] : [{ value, label: getLabel(value), count: 0 }],
-  );
-  return [...buckets, ...missing];
-};
-
 type SearchDialogProps = {
   open: boolean;
   onOpenChange: (open: boolean) => void;
   initialWorkspaceId?: string | undefined;
 };
-
-const initialSearchFilters = (
-  initialWorkspaceId: string | undefined,
-): SearchFilters => ({
-  editedByUserIds: [],
-  kinds: [],
-  mimeTypes: [],
-  types: [],
-  workspaceIds: initialWorkspaceId ? [initialWorkspaceId] : [],
-});
 
 export const SearchDialog = ({
   open,
@@ -1820,1224 +1665,4 @@ export const SearchDialog = ({
       </CommandDialogPopup>
     </CommandDialog>
   );
-};
-
-type SearchColumnResizeHandleProps = {
-  className?: string;
-  label: string;
-  /** Current width of the column this separator controls (px); the CSS
-   * default when the user has not dragged yet. */
-  value: number;
-  max: number;
-  min: number;
-  onResize: (clientX: number) => void;
-};
-
-const SEARCH_RESIZE_KEYBOARD_STEP_PX = 16;
-
-/**
- * Drag strip between two columns, mirroring the inspector pane's resize
- * handle. A 4px grab zone straddling the adjacent column border via
- * negative margins, so it adds no visible width of its own. Also a
- * focusable `separator`: arrow keys move it by a fixed step in the pressed
- * physical direction, feeding the same clamped `onResize` path as a drag
- * (a synthetic clientX offset from the handle's own position).
- */
-const SearchColumnResizeHandle = ({
-  className,
-  label,
-  value,
-  max,
-  min,
-  onResize,
-}: SearchColumnResizeHandleProps) => {
-  const isDraggingRef = useRef(false);
-
-  return (
-    <div
-      aria-label={label}
-      aria-orientation="vertical"
-      aria-valuemax={max}
-      aria-valuemin={min}
-      aria-valuenow={Math.round(value)}
-      className={cn(
-        "hover:bg-border active:bg-border focus-visible:bg-border focus-visible:ring-ring z-10 -mx-0.5 w-1 shrink-0 cursor-col-resize focus-visible:ring-1 focus-visible:outline-none",
-        className,
-      )}
-      onKeyDown={(event) => {
-        let step: number;
-        switch (event.key) {
-          case "ArrowLeft":
-            step = -SEARCH_RESIZE_KEYBOARD_STEP_PX;
-            break;
-          case "ArrowRight":
-            step = SEARCH_RESIZE_KEYBOARD_STEP_PX;
-            break;
-          default:
-            return;
-        }
-        event.preventDefault();
-        const rect = event.currentTarget.getBoundingClientRect();
-        onResize(rect.left + rect.width / 2 + step);
-      }}
-      onPointerDown={(event) => {
-        event.preventDefault();
-        isDraggingRef.current = true;
-        event.currentTarget.setPointerCapture(event.pointerId);
-      }}
-      onPointerMove={(event) => {
-        if (isDraggingRef.current) {
-          onResize(event.clientX);
-        }
-      }}
-      onPointerUp={() => {
-        isDraggingRef.current = false;
-      }}
-      onPointerCancel={() => {
-        isDraggingRef.current = false;
-      }}
-      role="separator"
-      tabIndex={0}
-    />
-  );
-};
-
-type SearchFooterHintProps = {
-  translationKey:
-    | "search.hintAskAI"
-    | "search.hintClose"
-    | "search.hintNavigate"
-    | "search.hintOpen";
-};
-
-const searchFooterKbd = (chunks: ReactNode) => (
-  <kbd className="border-border bg-muted rounded border px-1 py-0.5 text-[0.625rem] leading-none">
-    {chunks}
-  </kbd>
-);
-
-const SearchFooterHintText = ({ translationKey }: SearchFooterHintProps) => {
-  const t = useTranslations();
-  return <>{t.rich(translationKey, { kbd: searchFooterKbd })}</>;
-};
-
-// Keyboard hints mean nothing on touch viewports; below `sm` only the
-// footer's clickable controls (the Ask AI button) stay visible.
-const SearchFooterHint = ({ translationKey }: SearchFooterHintProps) => (
-  <span className="hidden whitespace-nowrap sm:inline">
-    <SearchFooterHintText translationKey={translationKey} />
-  </span>
-);
-
-type SearchPreviewPanelProps = {
-  dateVisibility?: "hide" | "show" | undefined;
-  hit: GlobalSearchHit;
-  locationModifierHeld?: boolean | undefined;
-  organizationId: string;
-  previewLocatorCandidates?: readonly string[] | undefined;
-  query: string;
-  userId: string;
-  onOpen: (hit: GlobalSearchHit) => void;
-};
-
-type RecentFilePreviewPanelProps = {
-  file: RecentFile;
-  locationModifierHeld?: boolean | undefined;
-  onOpen: () => void;
-  organizationId: string;
-  userId: string;
-};
-
-type RecentFilePreviewUnavailableProps = {
-  reason: "missing" | "request-error";
-};
-
-const RecentFilePreviewUnavailable = ({
-  reason,
-}: RecentFilePreviewUnavailableProps) => {
-  const t = useTranslations();
-  let message: string;
-  switch (reason) {
-    case "missing":
-      message = t("search.previewUnavailable");
-      break;
-    case "request-error":
-      message = t("common.somethingWentWrong");
-      break;
-    default: {
-      const exhaustive: never = reason;
-      return exhaustive;
-    }
-  }
-  return (
-    <aside className={SEARCH_PREVIEW_COLUMN_CLASS_NAME}>
-      <div className="text-muted-foreground flex flex-1 items-center justify-center px-5 text-center text-sm">
-        {message}
-      </div>
-    </aside>
-  );
-};
-
-const RecentFilePreviewPanel = ({
-  file,
-  locationModifierHeld = false,
-  onOpen,
-  organizationId,
-  userId,
-}: RecentFilePreviewPanelProps) => {
-  const isNativeDocument =
-    file.mimeType === PDF_MIME || file.mimeType === DOCX_MIME;
-  const fieldQuery = useQuery({
-    ...recentFilePreviewFieldOptions({
-      entityId: file.entityId,
-      fileFieldId: file.fileFieldId,
-      filePropertyId: file.filePropertyId,
-      mimeType: file.mimeType ?? null,
-      organizationId,
-      userId,
-      workspaceId: file.workspaceId,
-    }),
-    enabled: isNativeDocument,
-  });
-  if (isNativeDocument && fieldQuery.isPending) {
-    return (
-      <aside className={SEARCH_PREVIEW_COLUMN_CLASS_NAME}>
-        <NativeDocumentPreviewSkeleton />
-      </aside>
-    );
-  }
-  if (isNativeDocument && fieldQuery.isError) {
-    return <RecentFilePreviewUnavailable reason="request-error" />;
-  }
-  if (isNativeDocument && fieldQuery.data === null) {
-    return <RecentFilePreviewUnavailable reason="missing" />;
-  }
-
-  const resolvedFieldId = isNativeDocument
-    ? (fieldQuery.data ?? null)
-    : (file.fileFieldId ?? null);
-
-  const hit = getRecentFilePreviewHit(file, resolvedFieldId);
-
-  return (
-    <SearchPreviewPanel
-      dateVisibility={getRecentFilePreviewDateVisibility(file)}
-      hit={hit}
-      locationModifierHeld={locationModifierHeld}
-      onOpen={() => {
-        onOpen();
-      }}
-      organizationId={organizationId}
-      query=""
-      userId={userId}
-    />
-  );
-};
-
-const SearchPreviewPanel = ({
-  dateVisibility = "show",
-  hit,
-  locationModifierHeld = false,
-  organizationId,
-  previewLocatorCandidates = EMPTY_SEARCH_PREVIEW_LOCATOR_CANDIDATES,
-  query,
-  userId,
-  onOpen,
-}: SearchPreviewPanelProps) => {
-  const t = useTranslations();
-  const headingId = useId();
-
-  return (
-    <aside
-      aria-labelledby={headingId}
-      className={SEARCH_PREVIEW_COLUMN_CLASS_NAME}
-    >
-      <h2 className="sr-only" id={headingId}>
-        {t("common.preview")}
-      </h2>
-      <SearchPreviewContent
-        dateVisibility={dateVisibility}
-        hit={hit}
-        key={`${hit.type}:${hit.id}:${hit.updatedAt}:${normalizeSearchQuery(query)}`}
-        locationModifierHeld={locationModifierHeld}
-        onOpen={onOpen}
-        organizationId={organizationId}
-        previewLocatorCandidates={previewLocatorCandidates}
-        query={query}
-        userId={userId}
-      />
-    </aside>
-  );
-};
-
-type SearchPreviewContentProps = {
-  dateVisibility: "hide" | "show";
-  hit: GlobalSearchHit;
-  locationModifierHeld: boolean;
-  organizationId: string;
-  previewLocatorCandidates: readonly string[];
-  query: string;
-  userId: string;
-  onOpen: (hit: GlobalSearchHit) => void;
-};
-
-const SearchPreviewContent = ({
-  dateVisibility,
-  hit,
-  locationModifierHeld,
-  organizationId,
-  previewLocatorCandidates,
-  query,
-  userId,
-  onOpen,
-}: SearchPreviewContentProps) => {
-  const t = useTranslations();
-  const format = useFormatter();
-  const location =
-    hit.type === "contact" || hit.type === "case-law"
-      ? null
-      : hit.workspaceName;
-  // While Cmd/Ctrl is held, activation opens the hit's matter location, so
-  // the button says which matter it will open instead of the generic Open.
-  const opensLocation =
-    locationModifierHeld &&
-    location !== null &&
-    getEntityLocationRoute(hit) !== null;
-  const previewDate =
-    dateVisibility === "show" ? getSearchPreviewDate(hit) : null;
-  const formattedPreviewDate = previewDate
-    ? format.dateTime(new Date(previewDate.value), {
-        month: "short",
-        year: "numeric",
-        ...(previewDate.type === "calendar-date" ? { timeZone: "UTC" } : {}),
-      })
-    : null;
-
-  return (
-    <>
-      <div className="shrink-0 border-b px-4 py-3">
-        <div className="flex items-start gap-3">
-          <SearchHitIcon hit={hit} />
-          <div className="min-w-0 flex-1">
-            <BidiText as="p" className="line-clamp-2 text-sm font-medium">
-              {hit.title || hit.id}
-            </BidiText>
-            <p className="text-muted-foreground mt-0.5 truncate text-xs">
-              <BidiText>{t(KIND_TRANSLATION_KEYS[hit.type])}</BidiText>
-              {location && (
-                <>
-                  <span aria-hidden="true">{" · "}</span>
-                  <BidiText>{location}</BidiText>
-                </>
-              )}
-              {formattedPreviewDate && (
-                <>
-                  <span aria-hidden="true">{" · "}</span>
-                  <BidiText>{formattedPreviewDate}</BidiText>
-                </>
-              )}
-            </p>
-          </div>
-          <Button
-            className="min-w-0 shrink-0"
-            onClick={() => {
-              onOpen(hit);
-            }}
-            size="sm"
-            variant="outline"
-          >
-            {opensLocation ? (
-              <span className="max-w-48 truncate">
-                {t.rich("search.openMatter", {
-                  bdi: (chunks) => <BidiText>{chunks}</BidiText>,
-                  name: location,
-                })}
-              </span>
-            ) : (
-              t("common.open")
-            )}
-          </Button>
-        </div>
-      </div>
-      <SearchPreviewBody
-        hit={hit}
-        organizationId={organizationId}
-        previewLocatorCandidates={previewLocatorCandidates}
-        query={query}
-        userId={userId}
-      />
-    </>
-  );
-};
-
-type SearchPreviewBodyProps = Pick<
-  SearchPreviewContentProps,
-  "hit" | "organizationId" | "previewLocatorCandidates" | "query" | "userId"
->;
-
-const SearchPreviewBody = (props: SearchPreviewBodyProps) => {
-  const { hit } = props;
-  const searchText = useMemo(
-    () =>
-      getSearchHighlightText({
-        headline: hit.headline,
-        previewLocatorCandidates: props.previewLocatorCandidates,
-        query: props.query,
-      }),
-    [hit.headline, props.previewLocatorCandidates, props.query],
-  );
-  const emailPreviewTarget = getEmailSearchPreviewTarget(hit);
-  if (emailPreviewTarget) {
-    return (
-      <div className="min-h-0 flex-1 overflow-y-auto">
-        <Suspense fallback={<SearchTextPreviewSkeleton />}>
-          <EmailPreview
-            entityId={emailPreviewTarget.entityId}
-            fieldId={emailPreviewTarget.fieldId}
-            key={emailPreviewTarget.fieldId}
-            onOpenAttachment={ignoreSearchPreviewAttachment}
-            workspaceId={emailPreviewTarget.workspaceId}
-          />
-        </Suspense>
-      </div>
-    );
-  }
-
-  const nativePreviewTarget = getNativeSearchDocumentPreviewTarget(hit);
-  if (nativePreviewTarget) {
-    return (
-      <div className="min-h-0 flex-1 overflow-hidden">
-        <Suspense fallback={<NativeDocumentPreviewSkeleton />}>
-          <NativeDocumentPreview
-            fallback={<NativeDocumentPreviewSkeleton />}
-            key={`${hit.id}:${searchTextQueryKey(searchText)}`}
-            noMatchFallback={<SearchTextPreview {...props} />}
-            searchText={searchText}
-            target={nativePreviewTarget}
-          />
-        </Suspense>
-      </div>
-    );
-  }
-
-  return <SearchTextPreview {...props} />;
-};
-
-const NativeDocumentPreviewSkeleton = () => (
-  <div className="bg-muted/40 flex h-full justify-center overflow-hidden p-3">
-    <Skeleton className="h-[calc(100%+8rem)] w-full max-w-sm rounded-sm" />
-  </div>
-);
-
-const SearchTextPreview = ({
-  hit,
-  organizationId,
-  previewLocatorCandidates,
-  query,
-  userId,
-}: SearchPreviewBodyProps) => {
-  const t = useTranslations();
-  const searchText = useMemo(
-    () =>
-      getSearchHighlightText({
-        headline: hit.headline,
-        previewLocatorCandidates,
-        query,
-      }),
-    [hit.headline, previewLocatorCandidates, query],
-  );
-  const target = getSearchPreviewTarget(hit);
-  const { data, isError, isFetching, refetch } = useQuery(
-    searchPreviewOptions({
-      organizationId,
-      query,
-      resultId: target.resultId,
-      type: target.type,
-      updatedAt: hit.updatedAt,
-      userId,
-    }),
-  );
-  const authorizedData = selectAuthorizedSearchPreviewData({
-    data,
-    isError,
-  });
-  const renderContent = authorizedData
-    ? getSearchPreviewRenderContent(authorizedData)
-    : null;
-
-  return (
-    <div
-      className="min-h-0 flex-1 overflow-y-auto overscroll-contain px-5 py-5"
-      data-slot="search-preview-scroll-area"
-    >
-      {!isError && authorizedData === undefined && (
-        <div className="space-y-3">
-          <Skeleton className="h-4 w-full" />
-          <Skeleton className="h-4 w-11/12" />
-          <Skeleton className="h-4 w-4/5" />
-          <Skeleton className="mt-6 h-4 w-full" />
-          <Skeleton className="h-4 w-3/4" />
-        </div>
-      )}
-      {isError && (
-        <div className="flex min-h-48 flex-col items-center justify-center gap-3">
-          <p className="text-muted-foreground text-sm">
-            {t("common.somethingWentWrong")}
-          </p>
-          <Button
-            disabled={isFetching}
-            onClick={() => {
-              detached(refetch(), "search-dialog.refetch");
-            }}
-            size="sm"
-            variant="outline"
-          >
-            {t("common.retry")}
-          </Button>
-        </div>
-      )}
-      {renderContent?.type === "plain-text" && (
-        <div
-          className={SEARCH_PREVIEW_CONTENT_CLASS_NAME}
-          dir={contentDir(renderContent.directionText)}
-        >
-          {renderContent.text}
-        </div>
-      )}
-      {renderContent?.type === "chat-messages" && (
-        <Suspense fallback={<SearchTextPreviewSkeleton />}>
-          <ChatPreview
-            messages={renderContent.messages}
-            searchText={searchText}
-          />
-        </Suspense>
-      )}
-      {renderContent?.type === "highlighted-html" && (
-        <div
-          className={SEARCH_PREVIEW_CONTENT_CLASS_NAME}
-          dir={contentDir(renderContent.directionText)}
-          dangerouslySetInnerHTML={{
-            // safe-html: preview content is server-escaped before the trusted <mark> tags are inserted
-            __html: renderContent.html,
-          }}
-        />
-      )}
-    </div>
-  );
-};
-
-const SearchTextPreviewSkeleton = () => (
-  <div className="space-y-3">
-    <Skeleton className="h-12 w-4/5" />
-    <Skeleton className="ms-auto h-16 w-3/4" />
-    <Skeleton className="h-20 w-11/12" />
-  </div>
-);
-
-type SearchSummaryItemProps = {
-  summarizeMutation: UseMutationResult<
-    SearchSummaryData,
-    Error,
-    SearchAISummaryParams
-  >;
-  isOpeningChat: boolean;
-  onClick: () => void;
-  onOpenChat: () => void;
-  onCitationClick: (citationId: string) => void;
-};
-
-const SearchSummaryItem = ({
-  isOpeningChat,
-  summarizeMutation,
-  onClick,
-  onOpenChat,
-  onCitationClick,
-}: SearchSummaryItemProps) => {
-  const t = useTranslations();
-
-  const { isPending, isError, isSuccess, data } = summarizeMutation;
-
-  if (!isSuccess) {
-    let title = t("search.summaryAction");
-    let body = t("search.summaryPrompt");
-    if (isPending) {
-      title = t("search.summaryLoading");
-    } else if (isError) {
-      title = t("search.summaryError");
-      body = t("search.summaryRetry");
-    }
-
-    return (
-      <Button
-        className="mb-2 h-auto w-full items-start justify-start gap-3 rounded-md px-2.5 py-2.5 text-start whitespace-normal sm:h-auto"
-        disabled={isPending}
-        onClick={onClick}
-        variant={isError ? "destructive-outline" : "outline"}
-      >
-        <span className="bg-background text-foreground mt-0.5 flex size-6 shrink-0 items-center justify-center rounded-md border">
-          {isPending ? (
-            <LoaderIcon className="size-3.5 animate-spin" />
-          ) : (
-            <WandSparklesIcon className="size-3.5" />
-          )}
-        </span>
-        <span className="min-w-0 flex-1">
-          <span className="block truncate text-sm font-medium">{title}</span>
-          <span className="text-muted-foreground line-clamp-2 text-xs font-normal">
-            {body}
-          </span>
-        </span>
-      </Button>
-    );
-  }
-
-  return (
-    <div className="border-border bg-background mb-2 w-full rounded-md border px-2.5 py-2.5 text-start shadow-xs">
-      <div className="flex w-full items-start gap-3 text-start">
-        <span className="bg-background text-foreground mt-0.5 flex size-6 shrink-0 items-center justify-center rounded-md border">
-          <WandSparklesIcon className="size-3.5" />
-        </span>
-        <span className="min-w-0 flex-1">
-          <span className="block truncate text-sm font-medium">
-            {data.title}
-          </span>
-          <span className="text-muted-foreground block text-xs font-normal whitespace-pre-line">
-            <SummaryBody
-              citations={data.citations}
-              onCitationClick={onCitationClick}
-              text={data.summary}
-            />
-          </span>
-        </span>
-      </div>
-      <div className="border-border/70 mt-2 border-t pt-2">
-        <Button
-          className="h-auto gap-2 px-1.5 py-1 text-xs"
-          disabled={isOpeningChat}
-          onClick={onOpenChat}
-          size="sm"
-          variant="ghost"
-        >
-          {isOpeningChat ? (
-            <LoaderIcon className="size-3.5 animate-spin" />
-          ) : (
-            <MessageSquareIcon className="size-3.5" />
-          )}
-          {t("search.continueInChat")}
-        </Button>
-      </div>
-    </div>
-  );
-};
-
-type SummaryBodyProps = {
-  text: string;
-  citations: SearchSummaryData["citations"];
-  onCitationClick: (citationId: string) => void;
-};
-
-const CITATION_RE = /\[(?<number>\d+)\]/gu;
-
-const SummaryBody = ({
-  citations,
-  onCitationClick,
-  text,
-}: SummaryBodyProps) => {
-  const citationByNumber = new Map(
-    citations.map((citation) => [citation.number, citation]),
-  );
-  const parts: React.ReactNode[] = [];
-  let cursor = 0;
-
-  for (const match of text.matchAll(CITATION_RE)) {
-    const start = match.index;
-    const numberText = match.groups?.["number"];
-    const number = numberText ? Number(numberText) : Number.NaN;
-    const citation = citationByNumber.get(number);
-
-    if (start > cursor) {
-      parts.push(text.slice(cursor, start));
-    }
-
-    if (citation) {
-      parts.push(
-        <Tooltip
-          content={`${citation.title}\n${citation.reason}`}
-          key={`${citation.id}-${start}`}
-          layer="search-child"
-          render={
-            <button
-              className="text-foreground hover:bg-muted mx-0.5 rounded px-1 font-medium"
-              onClick={(event) => {
-                event.stopPropagation();
-                onCitationClick(citation.id);
-              }}
-              type="button"
-            >
-              [{citation.number}]
-            </button>
-          }
-        />,
-      );
-    } else {
-      parts.push(match[0]);
-    }
-
-    cursor = start + match[0].length;
-  }
-
-  if (cursor < text.length) {
-    parts.push(text.slice(cursor));
-  }
-
-  return parts;
-};
-
-type SearchRecentsProps = {
-  recentSearches: RecentSearch[];
-  recentFiles: RecentFile[];
-  onSearchClick: (recent: RecentSearch) => void;
-  onFileClick: (file: RecentFile) => void;
-  onFilePreview: (file: RecentFile) => void;
-  previewedFileId: string | null;
-};
-
-const SearchRecents = ({
-  recentSearches,
-  recentFiles,
-  onSearchClick,
-  onFileClick,
-  onFilePreview,
-  previewedFileId,
-}: SearchRecentsProps) => {
-  const t = useTranslations();
-  const hasRecents = recentSearches.length > 0 || recentFiles.length > 0;
-
-  // Scoped quick-create actions belong here once the action model is settled.
-  if (!hasRecents) {
-    return (
-      <div className="flex h-full items-center justify-center px-4 py-8">
-        <p className="text-muted-foreground text-sm">
-          {t("search.emptyState")}
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-5 px-4 py-4">
-      {recentSearches.length > 0 && (
-        <section>
-          <h3 className="text-muted-foreground mb-2 text-xs font-medium">
-            {t("search.recentSearches")}
-          </h3>
-          <div className="space-y-1">
-            {recentSearches.map((recent) => (
-              <Button
-                className="h-auto w-full justify-start gap-2 px-2 py-2 text-start text-sm"
-                data-search-empty-row=""
-                key={recent.query}
-                onClick={() => onSearchClick(recent)}
-                variant="ghost"
-              >
-                <HistoryIcon className="text-muted-foreground size-4 shrink-0" />
-                <span className="truncate">{recent.query}</span>
-              </Button>
-            ))}
-          </div>
-        </section>
-      )}
-
-      {recentFiles.length > 0 && (
-        <section>
-          <h3 className="text-muted-foreground mb-2 text-xs font-medium">
-            {t("search.recentlyOpenedFiles")}
-          </h3>
-          <div className="flex flex-col gap-y-1">
-            {recentFiles.map((file) => (
-              <Button
-                aria-current={
-                  previewedFileId === file.entityId ? "true" : undefined
-                }
-                className="h-auto! w-full justify-start gap-2 py-1 text-start text-sm"
-                data-previewing={previewedFileId === file.entityId}
-                data-search-empty-row=""
-                key={file.entityId}
-                onFocus={() => onFilePreview(file)}
-                onClick={() => {
-                  onFileClick(file);
-                }}
-                onPointerEnter={() => onFilePreview(file)}
-                variant={
-                  previewedFileId === file.entityId ? "secondary" : "ghost"
-                }
-              >
-                {file.mimeType ? (
-                  <DocumentIcon
-                    className="text-muted-foreground size-4 shrink-0"
-                    mimeType={file.mimeType}
-                  />
-                ) : (
-                  <FileTextIcon className="text-muted-foreground size-4 shrink-0" />
-                )}
-                <span className="min-w-0 flex-1">
-                  <BidiText as="span" className="block truncate">
-                    {file.title}
-                  </BidiText>
-                  <BidiText
-                    as="span"
-                    className="text-muted-foreground block truncate text-xs"
-                  >
-                    {file.workspaceName}
-                  </BidiText>
-                </span>
-              </Button>
-            ))}
-          </div>
-        </section>
-      )}
-    </div>
-  );
-};
-
-type FacetGroupProps = {
-  title: string;
-  buckets: {
-    value: string;
-    label?: string;
-    count: number;
-  }[];
-  selected: string[];
-  onChange: (value: string) => void;
-};
-
-type TimeFacetGroupProps = {
-  time: TimeFilter | undefined;
-  locale: string;
-  onPresetChange: (preset: TimePreset) => void;
-  onCustomChange: (range: { updatedFrom?: string; updatedTo?: string }) => void;
-  onClearCustom: () => void;
-};
-
-const TimeFacetGroup = ({
-  time,
-  locale,
-  onPresetChange,
-  onCustomChange,
-  onClearCustom,
-}: TimeFacetGroupProps) => {
-  const t = useTranslations();
-  const isCustom = time?.mode === "custom";
-  const customFromValue =
-    time?.mode === "custom" && time.updatedFrom
-      ? isoToDateInputValue(time.updatedFrom)
-      : null;
-  const customToValue =
-    time?.mode === "custom" && time.updatedTo
-      ? isoToDateInputValue(time.updatedTo)
-      : null;
-
-  const handleFromChange = (value: string | null) => {
-    if (!value) {
-      if (!customToValue) {
-        onClearCustom();
-        return;
-      }
-      onCustomChange({ updatedTo: dateInputToIsoEnd(customToValue) });
-      return;
-    }
-    onCustomChange({
-      updatedFrom: dateInputToIsoStart(value),
-      ...(customToValue && { updatedTo: dateInputToIsoEnd(customToValue) }),
-    });
-  };
-
-  const handleToChange = (value: string | null) => {
-    if (!value) {
-      if (!customFromValue) {
-        onClearCustom();
-        return;
-      }
-      onCustomChange({ updatedFrom: dateInputToIsoStart(customFromValue) });
-      return;
-    }
-    onCustomChange({
-      ...(customFromValue && {
-        updatedFrom: dateInputToIsoStart(customFromValue),
-      }),
-      updatedTo: dateInputToIsoEnd(value),
-    });
-  };
-
-  return (
-    <div>
-      <p className="text-muted-foreground mb-2 text-xs font-medium">
-        {t("search.updatedWithin")}
-      </p>
-      <div className="space-y-0.5">
-        {TIME_PRESETS.map((preset) => {
-          const isActive = time?.mode === "preset" && time.preset === preset;
-          return (
-            <Button
-              className="h-auto w-full justify-start gap-2 px-2 py-1 text-xs"
-              key={preset}
-              onClick={() => onPresetChange(preset)}
-              size="sm"
-              variant="ghost"
-            >
-              <Checkbox checked={isActive} tabIndex={-1} />
-              <span className="flex-1 truncate text-start">
-                {t(TIME_PRESET_TRANSLATION_KEYS[preset])}
-              </span>
-            </Button>
-          );
-        })}
-        <Button
-          className="h-auto w-full justify-start gap-2 px-2 py-1 text-xs"
-          onClick={() => {
-            if (isCustom) {
-              onClearCustom();
-            } else {
-              onCustomChange({});
-            }
-          }}
-          size="sm"
-          variant="ghost"
-        >
-          <Checkbox checked={isCustom} tabIndex={-1} />
-          <span className="flex-1 truncate text-start">
-            {t("search.timeFilterCustom")}
-          </span>
-        </Button>
-      </div>
-      {isCustom && (
-        <div className="mt-2 space-y-1 px-2">
-          <div>
-            <p className="text-muted-foreground text-[0.625rem] font-medium tracking-wide uppercase">
-              {t("search.dateFrom")}
-            </p>
-            <DatePickerPopover
-              layer="search-child"
-              locale={locale}
-              onChange={handleFromChange}
-              value={customFromValue}
-              {...(customToValue !== null && { maxDate: customToValue })}
-            />
-          </div>
-          <div>
-            <p className="text-muted-foreground text-[0.625rem] font-medium tracking-wide uppercase">
-              {t("search.dateTo")}
-            </p>
-            <DatePickerPopover
-              layer="search-child"
-              locale={locale}
-              onChange={handleToChange}
-              value={customToValue}
-              {...(customFromValue !== null && { minDate: customFromValue })}
-            />
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-const FacetGroup = ({
-  title,
-  buckets,
-  selected,
-  onChange,
-}: FacetGroupProps) => (
-  <div>
-    <p className="text-muted-foreground mb-2 text-xs font-medium">{title}</p>
-    <FacetBucketList
-      buckets={buckets}
-      onChange={onChange}
-      selected={selected}
-    />
-  </div>
-);
-
-type FacetBucket = { value: string; label?: string; count: number };
-
-const EMPTY_FACET_BUCKETS: readonly FacetBucket[] = [];
-const EMPTY_SEARCH_HITS: readonly GlobalSearchHit[] = [];
-const EMPTY_SEARCH_PREVIEW_LOCATOR_CANDIDATES: readonly string[] = [];
-
-type FacetBucketListProps = {
-  buckets: FacetBucket[];
-  selected: string[];
-  onChange: (value: string) => void;
-};
-
-const FacetBucketList = ({
-  buckets,
-  selected,
-  onChange,
-}: FacetBucketListProps) => {
-  const format = useFormatter();
-  return (
-    <div className="space-y-0.5">
-      {buckets.map((bucket) => (
-        <Button
-          className="h-auto w-full justify-start gap-2 px-2 py-1 text-xs"
-          key={bucket.value}
-          onClick={() => onChange(bucket.value)}
-          size="sm"
-          variant="ghost"
-        >
-          <Checkbox checked={selected.includes(bucket.value)} tabIndex={-1} />
-          <span className="flex-1 truncate text-start">
-            {bucket.label ?? bucket.value}
-          </span>
-          <span className="text-muted-foreground tabular-nums">
-            {format.number(bucket.count)}
-          </span>
-        </Button>
-      ))}
-    </div>
-  );
-};
-
-const FACET_SEARCH_DEBOUNCE_MS = 250;
-const FACET_SEARCH_LIMIT = 20;
-
-type SearchableFacetGroupProps = {
-  facet: SearchableFacet;
-  title: string;
-  defaultBuckets: readonly FacetBucket[];
-  selected: string[];
-  onChange: (value: string) => void;
-  searchParams: {
-    enabled: boolean;
-    organizationId: string;
-    userId: string;
-    query: string;
-    workspaceIds: string[];
-    types: GlobalSearchResultType[];
-    kinds: EntityKind[];
-    editedByUserIds: string[];
-    mimeTypes: string[];
-    updatedFrom?: string | undefined;
-    updatedTo?: string | undefined;
-  };
-  formatLabel?: (bucket: FacetBucket) => string;
-};
-
-const SearchableFacetGroup = ({
-  facet,
-  title,
-  defaultBuckets,
-  selected,
-  onChange,
-  searchParams,
-  formatLabel,
-}: SearchableFacetGroupProps) => {
-  const t = useTranslations();
-  const [search, setSearch] = useState("");
-  const [debouncedSearch, setDebouncedSearch] = useState("");
-  const debouncedSetSearch = useDebouncedCallback(
-    setDebouncedSearch,
-    FACET_SEARCH_DEBOUNCE_MS,
-  );
-  const labelCacheRef = useRef<Record<string, string>>({});
-  const isSearching = debouncedSearch.trim().length > 0;
-
-  const { data: searchData } = useQuery({
-    ...searchFacetOptions({
-      facet,
-      search: debouncedSearch,
-      limit: FACET_SEARCH_LIMIT,
-      ...searchParams,
-    }),
-    enabled: isSearching && searchParams.enabled,
-  });
-
-  const resolveLabel = (bucket: FacetBucket): string =>
-    formatLabel ? formatLabel(bucket) : (bucket.label ?? bucket.value);
-
-  // Refs are intentionally mutated during render — they don't trigger
-  // re-renders, and we want every label seen in the current render's
-  // buckets to be available when computing `buckets` below.
-  /* eslint-disable react/react-compiler -- deliberate render-time label cache: mutating labelCacheRef here makes every label seen this render available to the missingSelected lookup below; refs don't trigger re-renders so this is safe */
-  for (const bucket of defaultBuckets) {
-    labelCacheRef.current[bucket.value] = resolveLabel(bucket);
-  }
-  if (searchData) {
-    for (const bucket of searchData.buckets) {
-      labelCacheRef.current[bucket.value] = resolveLabel(bucket);
-    }
-  }
-  /* eslint-enable react/react-compiler */
-
-  const sourceBuckets =
-    isSearching && searchData ? searchData.buckets : defaultBuckets;
-
-  const visible: FacetBucket[] = sourceBuckets.map((bucket) => ({
-    value: bucket.value,
-    count: bucket.count,
-    label: resolveLabel(bucket),
-  }));
-  const present = new Set(visible.map((bucket) => bucket.value));
-  // eslint-disable-next-line react/react-compiler -- reads the deliberate render-time label cache mutated above; ref reads don't affect render correctness here
-  const missingSelected: FacetBucket[] = selected.flatMap((id) =>
-    present.has(id)
-      ? []
-      : [{ value: id, label: labelCacheRef.current[id] ?? id, count: 0 }],
-  );
-  const buckets: FacetBucket[] = [...visible, ...missingSelected];
-
-  if (buckets.length === 0 && !isSearching && searchData === undefined) {
-    return null;
-  }
-
-  return (
-    <div>
-      <p className="text-muted-foreground mb-2 text-xs font-medium">{title}</p>
-      <Input
-        className="mb-1.5 h-7 px-2 text-xs"
-        onChange={(e) => {
-          const value = e.target.value;
-          setSearch(value);
-          debouncedSetSearch(value);
-        }}
-        placeholder={t("common.search")}
-        value={search}
-      />
-      <FacetBucketList
-        buckets={buckets}
-        onChange={onChange}
-        selected={selected}
-      />
-    </div>
-  );
-};
-
-type SearchResultItemProps = {
-  hit: GlobalSearchHit;
-  index: number;
-  resultNumber: number;
-  onClick: (
-    hit: GlobalSearchHit,
-    options?: { locationModifier?: boolean },
-  ) => void;
-};
-
-const SearchResultItem = ({
-  hit,
-  index,
-  resultNumber,
-  onClick,
-}: SearchResultItemProps) => {
-  const t = useTranslations();
-  const format = useFormatter();
-  const date = new Date(hit.updatedAt);
-  const formatted = format.dateTime(date, {
-    month: "short",
-    year: "numeric",
-  });
-  let editorMeta: { image: string | null; name: string } | null = null;
-  if (hit.type === "document" && hit.lastEditedByName) {
-    editorMeta = {
-      image: hit.lastEditedByImage,
-      name: hit.lastEditedByName,
-    };
-  }
-  let meta: string;
-  if (hit.type === "contact") {
-    meta = t(KIND_TRANSLATION_KEYS[hit.type]);
-  } else if (hit.type === "case-law") {
-    meta = "";
-  } else if (hit.type === "matter") {
-    meta = compactMeta([hit.workspaceName, formatted]);
-  } else if (hit.type === "chat") {
-    meta = compactMeta([hit.workspaceName, formatted]);
-  } else {
-    const lastEditedByName =
-      hit.type === "document" ? null : hit.lastEditedByName;
-    meta = compactMeta([hit.workspaceName, formatted, lastEditedByName]);
-  }
-
-  return (
-    <CommandItem
-      className="h-auto w-full items-start justify-start gap-3 px-2 py-2 text-start whitespace-normal sm:h-auto"
-      index={index}
-      onClick={(event) =>
-        onClick(hit, {
-          locationModifier: event.metaKey || event.ctrlKey,
-        })
-      }
-      value={hit}
-    >
-      <SearchHitIcon hit={hit} />
-      <div className="min-w-0 flex-1">
-        <p className="truncate text-sm font-medium">{hit.title || hit.id}</p>
-        {(meta || editorMeta) && (
-          <div className="text-muted-foreground flex min-w-0 items-center gap-1 text-xs">
-            {meta ? <span className="min-w-0 truncate">{meta}</span> : null}
-            {editorMeta ? (
-              <>
-                {meta ? <span className="shrink-0">·</span> : null}
-                <UserIdentity
-                  as="span"
-                  avatarClassName="size-4 shrink-0 text-[0.5rem]"
-                  avatarFallbackClassName="text-[0.5rem]"
-                  className="gap-1.5"
-                  image={editorMeta.image}
-                  name={editorMeta.name}
-                  nameClassName="text-xs font-normal"
-                />
-              </>
-            ) : null}
-          </div>
-        )}
-        {hit.headline && (
-          <p
-            className="text-muted-foreground [&_mark]:bg-highlight [&_mark]:text-highlight-foreground mt-0.5 line-clamp-2 text-xs font-normal [&_mark]:font-medium"
-            dangerouslySetInnerHTML={{
-              // safe-html: server-escaped + <mark>-highlighted by escapeAndHighlight() in the global search mappers
-              __html: hit.headline,
-            }}
-          />
-        )}
-      </div>
-      <span className="text-foreground-subtle mt-0.5 shrink-0 px-1 text-xs tabular-nums">
-        {resultNumber}
-      </span>
-    </CommandItem>
-  );
-};
-
-const SearchHitIcon = ({ hit }: { hit: GlobalSearchHit }) => {
-  switch (hit.type) {
-    case "document":
-    case "folder":
-    case "task":
-    case "message":
-    case "link":
-      return (
-        <EntityKindIcon
-          className="text-muted-foreground mt-0.5 size-4 shrink-0"
-          kind={hit.type}
-          mimeType={hit.mimeType}
-        />
-      );
-    case "matter":
-      return (
-        <MatterIcon
-          className="mt-0.5 size-4 shrink-0"
-          matter={{ id: hit.workspaceId, color: hit.color }}
-        />
-      );
-    case "contact":
-    case "case-law":
-    case "chat": {
-      const Icon = NON_ENTITY_KIND_ICONS[hit.type];
-      return <Icon className="text-muted-foreground mt-0.5 size-4 shrink-0" />;
-    }
-    default: {
-      const exhaustive: never = hit;
-      return exhaustive;
-    }
-  }
 };
