+++ apps/web/src/features/chat/hooks/use-chat-session-send-queue.logic.test.ts
@@ -1,6 +1,7 @@
 import type { ChatClientState } from "@tanstack/ai-client";
 import { describe, expect, test } from "bun:test";
 
+import type { ChatUserMessageInput } from "@/features/chat/chat-runtime";
 import {
   createInitialSendQueueState,
   reduceSendQueue,
@@ -8,7 +9,6 @@ import {
   type QueuedChatEntry,
   type SendQueueState,
 } from "@/features/chat/hooks/use-chat-session-send-queue.logic";
-import type { ChatUserMessageInput } from "@/features/chat/queries";
 import { toSafeId } from "@/lib/safe-id";
 
 const CONVERSATION_ID = "conversation-1";
