+++ apps/web/src/features/chat/hooks/use-chat-thread-runtime.ts
@@ -1,10 +1,10 @@
 import { useQueryClient } from "@tanstack/react-query";
 
+import type { ChatThreadOptionsContext } from "@/features/chat/chat-query-contract";
+import type { ChatRuntime } from "@/features/chat/chat-runtime";
 import {
   acquireChatRuntime,
-  type ChatRuntime,
   type ChatThreadFetched,
-  type ChatThreadOptionsContext,
 } from "@/features/chat/queries";
 import type { ChatThreadRef } from "@/lib/chat-thread-ref";
