+++ scripts/react-compiler-bailouts.json
@@ -18,8 +18,8 @@
   "apps/web/src/components/pdf/peek/peek-pdf-viewer.tsx::PeekPrintButton": 1,
   "apps/web/src/components/pdf/peek/peek-pdf-viewer.tsx::PreparedPdfPrintButton": 1,
   "apps/web/src/components/pdf/versions-sidebar.tsx::VersionsSidebar": 0,
+  "apps/web/src/components/search-dialog-facets.tsx::SearchableFacetGroup": 0,
   "apps/web/src/components/search-dialog.tsx::SearchDialog": 4,
-  "apps/web/src/components/search-dialog.tsx::SearchableFacetGroup": 0,
   "apps/web/src/components/templates/template-form.tsx::TemplateForm": 11,
   "apps/web/src/components/workspaces/hooks/use-create-b-boxes.ts::useCreateBBoxes": 1,
   "apps/web/src/components/workspaces/hooks/use-sync-justifications.ts::useSyncJustificationChunks": 3,
